reorg finalizer and grpc and daemonClient
