+++ controllers/dnschaos/types.go
@@ -34,7 +34,9 @@ import (
 	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
 	"github.com/chaos-mesh/chaos-mesh/controllers/common"
 	"github.com/chaos-mesh/chaos-mesh/controllers/config"
+	"github.com/chaos-mesh/chaos-mesh/pkg/chaosdaemon/client"
 	"github.com/chaos-mesh/chaos-mesh/pkg/chaosdaemon/pb"
+	"github.com/chaos-mesh/chaos-mesh/pkg/finalizer"
 	"github.com/chaos-mesh/chaos-mesh/pkg/router"
 	ctx "github.com/chaos-mesh/chaos-mesh/pkg/router/context"
 	end "github.com/chaos-mesh/chaos-mesh/pkg/router/endpoint"
@@ -145,7 +147,7 @@ func (r *endpoint) cleanFinalizersAndRecover(ctx context.Context, chaos *v1alpha
 			}
 
 			r.Log.Info("Pod not found", "namespace", ns, "name", name)
-			chaos.Finalizers = utils.RemoveFromFinalizer(chaos.Finalizers, key)
+			chaos.Finalizers = finalizer.RemoveFromFinalizer(chaos.Finalizers, key)
 			continue
 		}
 
@@ -155,7 +157,7 @@ func (r *endpoint) cleanFinalizersAndRecover(ctx context.Context, chaos *v1alpha
 			continue
 		}
 
-		chaos.Finalizers = utils.RemoveFromFinalizer(chaos.Finalizers, key)
+		chaos.Finalizers = finalizer.RemoveFromFinalizer(chaos.Finalizers, key)
 	}
 
 	if chaos.Annotations[common.AnnotationCleanFinalizer] == common.AnnotationCleanFinalizerForced {
@@ -170,7 +172,7 @@ func (r *endpoint) cleanFinalizersAndRecover(ctx context.Context, chaos *v1alpha
 func (r *endpoint) recoverPod(ctx context.Context, pod *v1.Pod) error {
 	r.Log.Info("Try to recover pod", "namespace", pod.Namespace, "name", pod.Name)
 
-	daemonClient, err := utils.NewChaosDaemonClient(ctx, r.Client,
+	daemonClient, err := client.NewChaosDaemonClient(ctx, r.Client,
 		pod, config.ControllerCfg.ChaosDaemonPort)
 	if err != nil {
 		r.Log.Error(err, "get chaos daemon client")
@@ -210,7 +212,7 @@ func (r *endpoint) applyAllPods(ctx context.Context, pods []v1.Pod, chaos *v1alp
 			r.Log.Error(err, "get meta namespace key")
 			return err
 		}
-		chaos.Finalizers = utils.InsertFinalizer(chaos.Finalizers, key)
+		chaos.Finalizers = finalizer.InsertFinalizer(chaos.Finalizers, key)
 
 		g.Go(func() error {
 			return r.applyPod(ctx, pod, dnsServerIP)
@@ -227,7 +229,7 @@ func (r *endpoint) applyAllPods(ctx context.Context, pods []v1.Pod, chaos *v1alp
 func (r *endpoint) applyPod(ctx context.Context, pod *v1.Pod, dnsServerIP string) error {
 	r.Log.Info("Try to apply dns chaos", "namespace",
 		pod.Namespace, "name", pod.Name)
-	daemonClient, err := utils.NewChaosDaemonClient(ctx, r.Client,
+	daemonClient, err := client.NewChaosDaemonClient(ctx, r.Client,
 		pod, config.ControllerCfg.ChaosDaemonPort)
 	if err != nil {
 		r.Log.Error(err, "get chaos daemon client")
