+++ controllers/podiochaos/types.go
@@ -27,8 +27,8 @@ import (
 
 	"github.com/chaos-mesh/chaos-mesh/api/v1alpha1"
 	"github.com/chaos-mesh/chaos-mesh/controllers/config"
+	chaosDaemonClient "github.com/chaos-mesh/chaos-mesh/pkg/chaosdaemon/client"
 	"github.com/chaos-mesh/chaos-mesh/pkg/chaosdaemon/pb"
-	"github.com/chaos-mesh/chaos-mesh/pkg/utils"
 )
 
 // Handler applys podiochaos
@@ -52,7 +52,7 @@ func (h *Handler) Apply(ctx context.Context, chaos *v1alpha1.PodIoChaos) error {
 		return err
 	}
 
-	pbClient, err := utils.NewChaosDaemonClient(ctx, h, pod, config.ControllerCfg.ChaosDaemonPort)
+	pbClient, err := chaosDaemonClient.NewChaosDaemonClient(ctx, h, pod, config.ControllerCfg.ChaosDaemonPort)
 	if err != nil {
 		return err
 	}
