--- pkg/chaosfs/pb/injure.proto
@@ -1,31 +1,0 @@
-syntax = "proto3";
-
-package injure;
-
-import "google/protobuf/empty.proto";
-
-service Injure {
-  rpc Methods (google.protobuf.Empty) returns (Response) {}
-  rpc RecoverAll (google.protobuf.Empty) returns (google.protobuf.Empty) {}
-  rpc RecoverMethod (Request) returns (google.protobuf.Empty) {}
-  rpc SetFault(Request) returns (google.protobuf.Empty) {}
-  rpc SetFaultAll(Request) returns (google.protobuf.Empty) {}
-  rpc Injected (google.protobuf.Empty) returns (InjectedResponse) {}
-}
-
-message Request {
-  repeated string methods = 1;
-  uint32 errno = 2;
-  bool random = 3;
-  uint32 pct = 4;
-  string path = 5; // relative path (root is mountpoint)
-  uint32 delay = 6;
-}
-
-message Response {
-  repeated string methods = 1;
-}
-
-message InjectedResponse {
-    bool injected = 1;
-}
